refactor: move interfaces to separate files
