+++ src/artist/interfaces/artist.interface.ts
@@ -1,0 +1,5 @@
+export interface Artist {
+  id: string; // uuid v4
+  name: string;
+  grammy: boolean;
+}
