+++ src/user/interfaces/user-no-password.interface.ts
@@ -1,0 +1,7 @@
+export interface UserNoPassword {
+  id: string; // uuid v4
+  login: string;
+  version: number; // integer number, increments on update
+  createdAt: number; // timestamp of creation
+  updatedAt: number; // timestamp of last update
+}
