+++ src/user/interfaces/user.interface.ts
@@ -1,0 +1,5 @@
+import { UserNoPassword } from './user-no-password.interface';
+
+export interface User extends UserNoPassword {
+  password: string;
+}
