+++ src/utils/filterOutPassword.ts
@@ -1,4 +1,5 @@
-import { User, UserNoPassword } from 'src/user/user.service';
+import { UserNoPassword } from 'src/user/interfaces/user-no-password.interface';
+import { User } from 'src/user/interfaces/user.interface';
 
 export const filterOutPassword = (user: User): UserNoPassword => {
   const { password: _, ...filteredUser } = user;
